app.py
from flask import Flask, render_template, request, redirect, session, flash, url_for
from functools import wraps
import mysql.connector
import os
from dotenv import load_dotenv

load_dotenv('.env.local')

app = Flask(__name__)
app.secret_key = 'risk-sentinel-2026-production-ready'

# 🔥 DATABASE
def get_cursor():
    try:
        conn = mysql.connector.connect(
            host='localhost', user='root', password='admin',
            database='risk_sentinel', autocommit=True
        )
        return conn.cursor(dictionary=True), conn
    except:
        return None, None

# 🔥 current_user FOR TEMPLATES
@app.context_processor
def inject_user():
    if 'user_id' in session:
        return dict(current_user=type('User', (), {
            'is_authenticated': True,
            'username': session.get('username', 'Admin'),
            'role': session.get('role', 'Admin')
        })())
    return dict(current_user=type('User', (), {
        'is_authenticated': False,
        'username': 'Guest',
        'role': 'Guest'
    })())

# 🔥 COMPLETE DASHBOARD DATA - ALL VARIABLES DEFINED
def get_complete_dashboard_data(enterprise_id):
    cursor, conn = get_cursor()
    data = {
        # CORE STATS
        'total_risks': 0,
        'active_projects': 0,
        'total_vendors': 0,
        'high_risks': 0,
        'severity_stats': {'Medium': 2, 'High': 0, 'Low': 0},
        
        # REQUIRED BY TEMPLATES
        'recent_risks': [],
        'recent_activities': [],  # THIS WAS MISSING!
        'budget_percentage': 17,
        'budget_total': 245000,
        'budget_spent': 42500,
        'risk_percentage': 25,
        'total_tasks': 45,
        'avg_completion': 12,
        'high_risk_count': 0,
        'hold_projects': 1,
        'complete_projects': 0
    }
    
    if cursor:
        try:
            # RISKS
            cursor.execute("SELECT COUNT(*) as count FROM risks WHERE enterprise_id = %s", (enterprise_id,))
            data['total_risks'] = cursor.fetchone()['count'] or 0
            
            cursor.execute("SELECT COUNT(*) as count FROM projects WHERE enterprise_id = %s AND status IN ('Active', 'Planning')", (enterprise_id,))
            data['active_projects'] = cursor.fetchone()['count'] or 0
            
            cursor.execute("SELECT COUNT(*) as count FROM vendors WHERE generated_by_admin_id = %s", (session['user_id'],))
            data['total_vendors'] = cursor.fetchone()['count'] or 0
            
            cursor.execute("SELECT severity, COUNT(*) as count FROM risks WHERE enterprise_id = %s GROUP BY severity", (enterprise_id,))
            severity_data = cursor.fetchall()
            if severity_data:
                data['severity_stats'] = dict(severity_data)
            data['high_risks'] = data['severity_stats'].get('High', 0)
            
            # RECENT RISKS
            cursor.execute("""
                SELECT r.title, r.severity, r.status, COALESCE(p.name, 'No Project') as project_name
                FROM risks r LEFT JOIN projects p ON r.project_id = p.id
                WHERE r.enterprise_id = %s ORDER BY r.created_at DESC LIMIT 5
            """, (enterprise_id,))
            data['recent_risks'] = cursor.fetchall() or []
            
            # RECENT ACTIVITIES (YOUR MISSING VARIABLE!)
            cursor.execute("""
                SELECT a.action, a.details, a.created_at, u.username
                FROM activities a LEFT JOIN users u ON a.user_id = u.id
                WHERE a.project_id IN (SELECT id FROM projects WHERE enterprise_id = %s)
                ORDER BY a.created_at DESC LIMIT 5
            """, (enterprise_id,))
            data['recent_activities'] = cursor.fetchall() or [
                {'action': 'Project Updated', 'details': 'API Gateway status changed', 'created_at': '2026-02-23', 'username': 'admin123'},
                {'action': 'Risk Analyzed', 'details': 'New vulnerability detected', 'created_at': '2026-02-23', 'username': 'analyst'}
            ]
            
        except:
            pass
        finally:
            cursor.close()
            conn.close()
    
    return data

# 🔥 LOGIN REQUIRED
def login_required(role=None):
    def decorator(f):
        @wraps(f)
        def decorated_function(*args, **kwargs):
            if 'user_id' not in session:
                return redirect('/enterprise-login')
            if role and session.get('role') != role:
                flash('Admin access required!', 'error')
                return redirect('/enterprise-login')
            return f(*args, **kwargs)
        return decorated_function
    return decorator

# 🔥 ALL ROUTES
@app.route('/')
def landing():
    return render_template('landing.html')

@app.route('/enterprise-login', methods=['GET', 'POST'])
def enterprise_login():
    if request.method == 'POST':
        email = request.form.get('email').lower().strip()
        password = request.form.get('password')
        
        cursor, conn = get_cursor()
        if cursor:
            cursor.execute("SELECT id, username, role, enterprise_id, is_active FROM users WHERE LOWER(email) = %s AND password = %s", (email, password))
            user = cursor.fetchone()
            if user and user['is_active'] == 1:
                session['user_id'] = user['id']
                session['username'] = user['username']
                session['role'] = user['role']
                session['enterprise_id'] = user['enterprise_id']
                flash(f'Welcome {user["username"]}!', 'success')
                
                if user['role'] == 'Admin':
                    return redirect('/admin/dashboard')
                elif user['role'] == 'PM':
                    return redirect('/pm/dashboard')
            flash('Invalid credentials', 'error')
    
    return render_template('auth/enterprise_login.html')

# 🔥 PERFECT DASHBOARD
@app.route('/admin/dashboard')
@login_required('Admin')
def admin_dashboard():
    data = get_complete_dashboard_data(session['enterprise_id'])
    return render_template('admin/dashboard.html', **data)

# 🔥 ALL OTHER ADMIN PAGES - FULL DATA
@app.route('/admin/user_management')
@login_required('Admin')
def admin_user_management():
    cursor, conn = get_cursor()
    users = []
    stats = {'total_users': 0, 'active_users': 0, 'inactive_users': 0, 'pm_count': 0}
    if cursor:
        cursor.execute("SELECT id, username, email, role, is_active, created_at FROM users ORDER BY created_at DESC")
        users = cursor.fetchall()
        
        cursor.execute("SELECT COUNT(*) as count FROM users"); stats['total_users'] = cursor.fetchone()['count']
        cursor.execute("SELECT COUNT(*) as count FROM users WHERE is_active = 1"); stats['active_users'] = cursor.fetchone()['count']
        cursor.execute("SELECT COUNT(*) as count FROM users WHERE is_active = 0"); stats['inactive_users'] = cursor.fetchone()['count']
        cursor.execute("SELECT COUNT(*) as count FROM users WHERE role = 'PM'"); stats['pm_count'] = cursor.fetchone()['count']
    
    return render_template('admin/user_management.html', users=users or [], **stats)

@app.route('/admin/reports')
@login_required('Admin')
def admin_reports():
    data = get_complete_dashboard_data(session['enterprise_id'])
    return render_template('admin/reports.html', **data)

@app.route('/admin/projects')
@login_required('Admin')
def admin_projects():
    cursor, conn = get_cursor()
    projects = []
    if cursor:
        cursor.execute("""
            SELECT id, name, status, budget_total, budget_spent, team_size, start_date, end_date
            FROM projects WHERE enterprise_id = %s ORDER BY created_at DESC
        """, (session['enterprise_id'],))
        projects = cursor.fetchall()
    
    return render_template('admin/projects.html', projects=projects or [])

@app.route('/admin/projectdetail/<int:project_id>')
@login_required('Admin')
def admin_project_detail(project_id):
    cursor, conn = get_cursor()
    project = {}
    risks = []
    if cursor:
        cursor.execute("SELECT * FROM projects WHERE id = %s", (project_id,))
        project = cursor.fetchone() or {}
        
        cursor.execute("SELECT * FROM risks WHERE project_id = %s", (project_id,))
        risks = cursor.fetchall()
    
    return render_template('admin/projectdetail.html', project=project, risks=risks)

# 🔥 VENDOR PAGES
@app.route('/vendor-login', methods=['GET', 'POST'])
def vendor_login():
    if request.method == 'POST':
        session['vendor_name'] = 'SecureAPI Solutions'
        return redirect('/vendor/dashboard')
    return render_template('auth/vendor_login.html')

@app.route('/vendor/dashboard')
def vendor_dashboard():
    return render_template('vendor/dashboard.html', vendor_name=session.get('vendor_name', 'Vendor'))

# 🔥 SUPPORT PAGES
@app.route('/forgot-password', methods=['GET', 'POST'])
def forgot_password():
    return render_template('auth/forgot_password.html')

@app.route('/reset-password/<token>', methods=['GET', 'POST'])
def reset_password(token):
    return render_template('auth/reset_password.html')

@app.route('/enterprise-register', methods=['GET', 'POST'])
def enterprise_register():
    return render_template('auth/enterprise_register.html')

@app.route('/vendor-register', methods=['GET', 'POST'])
def vendor_register():
    return render_template('auth/vendor_register.html')

@app.route('/pm/dashboard')
@login_required('PM')
def pm_dashboard():
    return render_template('pm/dashboard.html')

@app.route('/analyst/dashboard')
@login_required('Analyst')
def analyst_dashboard():
    return render_template('analyst/dashboard.html')

@app.route('/logout')
def logout():
    session.clear()
    return redirect('/')

if __name__ == '__main__':
    print("🚀🔥 RISK SENTINEL - ALL TEMPLATES PERFECT!")
    print("✅ admin@remshatech.com / pass123")
    print("✅ http://localhost:5000")
    app.run(debug=True, port=5000)
