requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
Werkzeug==3.0.3
Flask-Limiter==3.8.0
cryptography==42.0.5
itsdangerous==2.2.0
